+++ src/main/java/org/opentripplanner/graph_builder/module/DirectTransferGenerator.java
@@ -8,7 +8,7 @@ import org.opentripplanner.model.SimpleTransfer;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.GraphIndex;
-import org.opentripplanner.routing.graph_finder.StopAtDistance;
+import org.opentripplanner.routing.graphfinder.StopAtDistance;
 import org.opentripplanner.routing.vertextype.TransitStopVertex;
 import org.opentripplanner.util.ProgressTracker;
 import org.slf4j.Logger;
